Factor out VeProcess for cleanliness
We added so many new methods now
