+++ src/test/scala/com/nec/ve/WithVeProcess.scala
@@ -7,7 +7,7 @@ import org.scalatest.{BeforeAndAfterAll, Suite}
 trait WithVeProcess extends BeforeAndAfterAll { this: Suite =>
 
   implicit def source: VeColVectorSource = VeColVectorSource(s"VE Tests")
-  implicit def veProcess: VeProcess = VeProcess.WrappingVeo(proc, source, VeProcessMetrics.NoOp)
+  implicit def veProcess: VeProcess = VeProcess.wrappingVeo(proc, source, VeProcessMetrics.NoOp)
 
   private var initialized = false
